Removes noisy child classes for gt, gte etc
Previously we had implemented an abstract class. We decided it would be
cleaner to pass args into a constructor and create simpler anonymous
classes.

This mirrors the functionality of and, or, etc
